requirements.txt
langchain
langchain-chroma
langchain-text-splitters
langchain-groq
langchain-huggingface
sentence-transformers
chromadb
beautifulsoup4
requests
pdfplumber
streamlit
tiktoken
python-dotenv